common/common-HTMLOptionsCollection.html

<!DOCTYPE html>
<html xmlns:h="http://www.w3.org/2000/xhtml">
<head>
	<meta charset="utf-8">

	<title id='title'>HTMLOptionsCollection 2.8.3.2</title>
	<script src="/resources/testharnessreport.js"></script>	
	<script src="/resources/testharness.js"></script>

	<div id="log"></div>
    <select id="selly">
        <option>1</option>
        <option>2</option>
        <option>3</option>
        <option>4</option>
    </select>

    	<script>
    	    (function () {
    	        //test asserts go here
    	        var selly = document.getElementById('selly');
    	        var initval = selly.length;
    	        var midval = 0;
    	        var firecount = 0;
    	        selly.addEventListener('DOMNodeInserted', firecounter)

    	        function firecounter() {
    	            firecount++;
    	        }

    	        test(function () {
    	            assert_equals(selly.length, 4);
    	        }, "On getting, the length attribute must return the number of nodes represented by the collection.");

    	        selly.length = 7;
    	        test(function () {
    	            assert_equals(selly.length, 7);
    	        }, "Changing the length adds new nodes");

    	        test(function () {
    	            midval = selly.length;
    	            assert_equals(midval - initval, 3);
    	        }, "The number of new nodes = new length minus old length");

    	        var elarray = [];
    	        for (var i = 0; i < selly.length; i++) {
    	            elarray.push(selly[i].value);
    	        }

    	        test(function () {
    	            assert_array_equals(elarray, ["1", "2", "3", "4", "", "", ""]);
    	        }, "New nodes have no value");

    	        test(function () {
    	            assert_equals(firecount, 3);
    	        }, "Changing the length & adding new nodes fires DOMNodeInserted on the select once for each new element");

    	        selly.length = 7;
    	        test(function () {
    	            assert_equals(selly.length, 7);
    	        }, "Setting a length equal to existing length changes nothing");


    	        selly.length = 4;
    	        test(function () {
    	            assert_equals(selly[6], undefined);
    	        }, "Setting a length lower than the old length trims nodes from the end (checking that previously set node is now undefined)");

    	        test(function () {
    	            var newdiff = midval - selly.length;
    	            assert_equals(newdiff, 3);
    	        }, "Number of nodes trimmed from the end is equal to old length - new length");

    	    })();
	</script>
